pred presentation

+++ app/src/main/java/com/example/jusanbookingapp/domain/use_cases/GetRoomReservationsByDateUseCase.kt
@@ -4,7 +4,7 @@ import com.example.jusanbookingapp.domain.models.Reservation
 import com.example.jusanbookingapp.domain.repositories.ReservationRepository
 
 class GetRoomReservationsByDateUseCase(private val reservationRepository: ReservationRepository) {
-    suspend operator fun invoke(roomNumber : String, startTime : Int, endTime : Int) : List<Reservation> {
+    suspend operator fun invoke(roomNumber : String, startTime : Long, endTime : Long) : List<Reservation> {
         return reservationRepository.getRoomReservationsByDate(roomNumber = roomNumber, startTime = startTime, endTime = endTime)
     }
 }
